ai-stock-discovery-engine/requirements.txt
# Core dependencies - Python 3.9+ compatible
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
python-dotenv>=1.0.0
tweepy>=4.14.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0
finnhub-python>=2.4.0
alpha-vantage>=2.3.1
polygon-api-client>=1.12.0
newsapi-python>=0.2.6
schedule>=1.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6

# TensorFlow compatibility (if needed)
tensorflow>=2.13.0
jax>=0.4.0
typing-extensions>=4.0.0
